+++ src/main/resources/db/migration/V3__createstadiumstable.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS stadiums (
-                                        id BIGINT PRIMARY KEY,
+                                        id SERIAL PRIMARY KEY,
                                         name VARCHAR(255) NOT NULL,
                                         location VARCHAR(255) NOT NULL,
                                         capacity INT NOT NULL
